refactor(console): workspace controller (wave 2)
Extracts ConsoleWorkspaceController (tldw_chatbook/UI/Console_Modules/
workspace.py) from ChatScreen: the largest cluster in wave 2 -- 35 methods,
~1,341 lines of method bodies covering workspace policy context, the
switcher/rename/archive/create lifecycle, the workspace RAG-scope picker,
workspace-scoped Console session activation, resuming a persisted
conversation into a native Console session, and the workspace conversation
search rows.

Follows the wave-1 ConsoleDictationController binding rule: framework
services (run_worker/push_screen/call_after_refresh) live-read via
@property; sibling clusters' own state reach back the same way
(_pending_console_launch_context, _console_agent_drilldown_run_id); every
other screen dependency is a named, late-bound constructor callable so
every moved method body is a byte-for-byte copy of the pre-extraction
source. Two of this cluster's own original members
(_sync_console_workspace_context, _focus_console_workspace_conversation_
search) touch the DOM and stay on ChatScreen per the "zero DOM in the
controller" rule, reached via named callables rather than screen
properties since they belong to this cluster, not a sibling one.

ChatScreen keeps six get/set proxy properties for
_console_workspace_conversation_query/_search_timer/_search_token/
_search_rows/_search_total/_search_error (not _workspace_id, which is
purely internal to the moved _with_console_workspace_conversation_
section) so on_console_workspace_conversation_search_changed (a sibling
conversation-browser handler that only mirrors three of them) and
on_button_pressed's workspace-search branches needed no change.

Characterisation (previous commit) plus the existing 209+166+84-test
workspace/resume/search/browser suite all still pass unchanged; pyflakes
before/after comparison on chat_screen.py matches the pre-move baseline
exactly (only the 13 now-workspace.py-owned imports removed); all 50
external call sites and 8 test-helper files referencing the moved methods
were retargeted to console._workspace.<method>.
