+++ tldw_chatbook/UI/Screens/chat_screen.py
@@ -67,6 +67,7 @@ from ..Console_Modules.dictation import (
 )
 from ..Console_Modules.left_rail import ConsoleLeftRail
 from ..Console_Modules.right_rail import ConsoleInspectorRail
+from ..Console_Modules.workspace import ConsoleWorkspaceController
 from ...Chat.chat_persistence_service import ChatPersistenceService
 from ...Chat.citation_trace_repository import ActiveCitationTraceState
 from ...Chat.console_chat_controller import ConsoleChatController
@@ -172,8 +173,6 @@ from ...Chat.console_chat_models import (
     ConsoleRunStatus,
     ConsoleVariant,
     ConsoleVariantSet,
-    ConsoleWorkspaceContext,
-    ConsoleStagedSource,
     MessageAttachment,
     derive_console_session_title,
 )
@@ -293,7 +292,6 @@ from ...Chat.console_display_state import (
     console_prompted_evidence_text,
     console_prompted_source_count,
     console_staged_source_count,
-    evidence_bundle_from_launch,
 )
 from ...Chat.console_onboarding_state import (
     ConsoleSetupCardState,
@@ -430,8 +428,6 @@ from ...Widgets.Console import (
     ConsoleStagedEvidenceStrip,
     ConsoleTranscript,
     ConsoleWorkspaceContextTray,
-    ConsoleWorkspaceRenameModal,
-    ConsoleWorkspaceSwitcherModal,
 )
 from ...Widgets.confirmation_dialog import ConfirmationDialog
 from ...Widgets.Console.console_image_viewer_modal import (
@@ -535,24 +531,14 @@ from ...Widgets.Console.console_rewind_modal import (
 from ...Widgets.Console.console_workspace_details import ConsoleWorkspaceDetailsTray
 from ...Widgets.Console.console_workbench_state import build_console_workbench_state
 from ...Workspaces.display_state import (
-    CONSOLE_WORKSPACE_CONVERSATION_RESULT_LIMIT,
     ConsoleWorkspaceConversationRow,
-    ConsoleWorkspaceConversationSectionState,
     ConsoleWorkspaceContextState,
-    build_console_workspace_state,
-    console_workspace_conversation_result_copy,
-)
-from ...Workspaces.registry_service import (
-    WorkspaceNotFound,
-    WorkspaceRegistryServiceError,
-    next_local_workspace_identity,
 )
 from ...Workspaces import (
     CONSOLE_CONVERSATION_BROWSER_RESULT_LIMIT,
     ConsoleConversationBrowserInputRow,
     ConsoleConversationBrowserRow,
     DEFAULT_WORKSPACE_ID,
-    WorkspaceRecord,
     build_console_conversation_browser_state,
     console_persisted_row_updated_sort,
 )
@@ -2559,7 +2545,7 @@ class ChatScreen(BaseAppScreen):
         if controller.store.active_session_id != session_id:
             self._capture_console_draft_switch_snapshot()
             self._note_console_follow_intent()
-            self._set_active_workspace_for_console_session(session_id)
+            self._workspace._set_active_workspace_for_console_session(session_id)
             controller.switch_session(session_id)
             # Finding C: a sub-agent drill-in is scoped to the conversation
             # active when the user drilled in -- clear it immediately on
@@ -2615,7 +2601,7 @@ class ChatScreen(BaseAppScreen):
             await self._activate_native_console_session(entry.native_session_id)
             return
         if entry.conversation_id:
-            resumed = await self._resume_console_workspace_conversation(
+            resumed = await self._workspace._resume_console_workspace_conversation(
                 entry.conversation_id,
                 target_scope_type=entry.scope_type or None,
                 target_workspace_id=entry.workspace_id,
@@ -2673,7 +2659,7 @@ class ChatScreen(BaseAppScreen):
     def on_console_change_workspace(self, event: Button.Pressed) -> None:
         """Open the active Console workspace switcher."""
         event.stop()
-        self._open_console_workspace_switcher()
+        self._workspace._open_console_workspace_switcher()
 
     @on(Button.Pressed, "#console-new-temporary-tab")
     def on_console_new_temporary_tab(self, event: Button.Pressed) -> None:
@@ -2689,250 +2675,17 @@ class ChatScreen(BaseAppScreen):
 
     def action_open_console_workspace_switcher(self) -> None:
         """Open the workspace switcher (Alt+W / command palette, TASK-722)."""
-        self._open_console_workspace_switcher()
-
-    def _open_console_workspace_switcher(self) -> None:
-        """Open the active Console workspace switcher."""
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is None:
-            self.app_instance.notify(
-                "Workspace service is not ready.", severity="warning"
-            )
-            return
-        try:
-            workspaces = tuple(registry_service.list_workspaces())
-            active_workspace = registry_service.get_active_workspace()
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Unable to open Console workspace switcher"
-            )
-            self.app_instance.notify(
-                "Workspace registry could not be read.",
-                severity="error",
-            )
-            return
-        if not workspaces:
-            self.app_instance.notify(
-                "Create one with the rail's New button or in Settings > Workspaces.",
-                severity="warning",
-            )
-            return
-
-        active_workspace_id = (
-            active_workspace.workspace_id if active_workspace is not None else None
-        )
-
-        def _switch_to(workspace_id: str) -> None:
-            try:
-                registry_service.set_active_workspace(workspace_id)
-            except Exception:
-                logger.opt(exception=True).warning(
-                    "Unable to switch Console workspace",
-                )
-                self.app_instance.notify(
-                    "Workspace could not be selected.",
-                    severity="error",
-                )
-                return
-            self._sync_console_chat_core_state()
-            self._activate_console_session_for_workspace(workspace_id)
-            self._sync_console_workspace_context()
-            self.run_worker(
-                self._sync_native_console_chat_ui(),
-                exclusive=True,
-                group="console-sync",
-            )
-
-        def _apply_workspace_switch(
-            result: tuple[str, str] | None,
-        ) -> None:
-            if not result:
-                return
-            action, workspace_id = result
-            if action == "switch":
-                _switch_to(workspace_id)
-            elif action == "rename":
-                self._open_console_workspace_rename(workspace_id)
-            elif action == "archive":
-                self._confirm_console_workspace_archive(workspace_id)
-
-        self.app.push_screen(
-            ConsoleWorkspaceSwitcherModal(
-                workspaces=workspaces,
-                active_workspace_id=active_workspace_id,
-            ),
-            callback=_apply_workspace_switch,
-        )
-
-    def _open_console_workspace_rename(self, workspace_id: str) -> None:
-        """Prompt for and apply a new workspace name (TASK-714)."""
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is None:
-            self.app_instance.notify(
-                "Workspace service is not ready.", severity="warning"
-            )
-            return
-        record = registry_service.get_workspace(workspace_id)
-        if record is None:
-            self.app_instance.notify(
-                "Workspace is no longer available.", severity="warning"
-            )
-            return
-
-        def _apply_rename(new_name: str | None) -> None:
-            if not new_name:
-                return
-            try:
-                renamed = registry_service.rename_workspace(workspace_id, new_name)
-            except WorkspaceRegistryServiceError as exc:
-                self.app_instance.notify(str(exc), severity="warning")
-                return
-            except Exception:
-                logger.opt(exception=True).warning("Unable to rename Console workspace")
-                self.app_instance.notify(
-                    "Workspace could not be renamed.", severity="error"
-                )
-                return
-            self._sync_console_chat_core_state()
-            self._sync_console_workspace_context()
-            self.run_worker(
-                self._sync_native_console_chat_ui(),
-                exclusive=True,
-                group="console-sync",
-            )
-            self.app_instance.notify(
-                f"Renamed workspace to {renamed.name}.", severity="information"
-            )
-
-        self.app.push_screen(
-            ConsoleWorkspaceRenameModal(current_name=record.name),
-            callback=_apply_rename,
-        )
-
-    def _confirm_console_workspace_archive(self, workspace_id: str) -> None:
-        """Confirm and archive a workspace (TASK-714)."""
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is None:
-            self.app_instance.notify(
-                "Workspace service is not ready.", severity="warning"
-            )
-            return
-        record = registry_service.get_workspace(workspace_id)
-        if record is None:
-            self.app_instance.notify(
-                "Workspace is no longer available.", severity="warning"
-            )
-            return
-        was_active = bool(record.active)
-
-        # ConfirmationDialog awaits its confirm callback, so this must be a
-        # coroutine function.
-        async def _archive() -> None:
-            try:
-                registry_service.archive_workspace(workspace_id)
-            except WorkspaceRegistryServiceError as exc:
-                self.app_instance.notify(str(exc), severity="warning")
-                return
-            except Exception:
-                logger.opt(exception=True).warning(
-                    "Unable to archive Console workspace"
-                )
-                self.app_instance.notify(
-                    "Workspace could not be archived.", severity="error"
-                )
-                return
-            self._sync_console_chat_core_state()
-            if was_active:
-                self._activate_console_session_for_workspace(DEFAULT_WORKSPACE_ID)
-            self._sync_console_workspace_context()
-            self.run_worker(
-                self._sync_native_console_chat_ui(),
-                exclusive=True,
-                group="console-sync",
-            )
-            suffix = " Console switched to the Default workspace." if was_active else ""
-            self.app_instance.notify(
-                f"Archived {record.name}. Its conversations stay saved in "
-                f"Library.{suffix}",
-                severity="information",
-            )
-
-        self.app.push_screen(
-            ConfirmationDialog(
-                title="Archive workspace?",
-                message=(
-                    f"Archive {record.name}? Its conversations stay saved and "
-                    "remain visible in Library; the workspace disappears from "
-                    "the switcher and the Console browser."
-                ),
-                confirm_label="Archive",
-                confirm_callback=_archive,
-            )
-        )
+        self._workspace._open_console_workspace_switcher()
 
     @on(Button.Pressed, "#console-new-workspace")
     def on_console_new_workspace(self, event: Button.Pressed) -> None:
         """Create a new local workspace from the Console rail and activate it."""
         event.stop()
-        self._create_console_workspace()
+        self._workspace._create_console_workspace()
 
     def action_new_console_workspace(self) -> None:
         """Create a local workspace from the command palette (TASK-722)."""
-        self._create_console_workspace()
-
-    def _create_console_workspace(self) -> None:
-        """Create a new local workspace and activate it."""
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is None:
-            self.app_instance.notify(
-                "Workspace service is not ready.", severity="warning"
-            )
-            return
-        try:
-            workspace_id, workspace_name = next_local_workspace_identity(
-                registry_service
-            )
-            registry_service.create_workspace(
-                workspace_id=workspace_id,
-                name=workspace_name,
-                description="Local workspace created from Console.",
-            )
-            registry_service.set_active_workspace(workspace_id)
-        except WorkspaceRegistryServiceError:
-            logger.opt(exception=True).warning("Unable to create Console workspace")
-            self.app_instance.notify(
-                "Workspace could not be created.", severity="error"
-            )
-            return
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Unexpected error creating Console workspace"
-            )
-            self.app_instance.notify(
-                "Workspace could not be created.", severity="error"
-            )
-            return
-        self._sync_console_chat_core_state()
-        self._activate_console_session_for_workspace(workspace_id)
-        self._sync_console_workspace_context()
-        self.run_worker(
-            self._sync_native_console_chat_ui(), exclusive=True, group="console-sync"
-        )
-        # TASK-713: creation also activates the workspace and opens a tab;
-        # without a notification the whole sequence is invisible when the
-        # Workspace status row is scrolled out of view.
-        self.app_instance.notify(
-            f"Created {workspace_name} and switched Console to it.",
-            severity="information",
-        )
+        self._workspace._create_console_workspace()
 
     @on(Button.Pressed, "#console-workspace-rag-scope-open")
     def on_console_workspace_rag_scope_open(self, event: Button.Pressed) -> None:
@@ -2944,137 +2697,10 @@ class ChatScreen(BaseAppScreen):
         """
         event.stop()
         self.run_worker(
-            self._open_console_workspace_scope_picker(),
+            self._workspace._open_console_workspace_scope_picker(),
             exclusive=True,
             group="console-workspace-scope-open",
         )
-
-    async def _open_console_workspace_scope_picker(self) -> None:
-        """Open the RAG retrieval-scope picker for the ACTIVE workspace.
-
-        Task-13 workspace entry point (design spec section 4, "Workspace
-        entry: Scope button beside the workspace row in the Session area").
-        ``universe=None`` -- the workspace picker offers the full library,
-        unlike the conversation-target picker, which restricts to the
-        workspace's own items once one is set (D3, see
-        ``_open_console_retrieval_scope_picker``).
-
-        Only a REAL registry workspace can be scoped -- gated the same way
-        the mounted button itself is gated
-        (``ConsoleWorkspaceContextState.rag_scope_enabled``): the built-in
-        Default workspace has a real ``workspace_id`` row
-        (``DEFAULT_WORKSPACE_ID``) once ``ensure_default_workspace`` has
-        run, so it IS scopable; only the "Local Default"/error/no-registry
-        sentinel states (no real workspace row at all) are refused here.
-        """
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is None:
-            self.app_instance.notify(
-                "Workspace service is not ready.", severity="warning"
-            )
-            return
-        try:
-            active_workspace = registry_service.get_active_workspace()
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Unable to read active workspace for the scope picker"
-            )
-            self.app_instance.notify(
-                "Workspace registry could not be read.", severity="error"
-            )
-            return
-        if active_workspace is None:
-            self.app_instance.notify(
-                "Create or select a workspace before setting a RAG scope.",
-                severity="warning",
-            )
-            return
-        workspace_id = active_workspace.workspace_id
-
-        try:
-            initial = await self._read_console_workspace_scope(
-                registry_service, workspace_id
-            )
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Unable to read workspace scope for {}", workspace_id
-            )
-            initial = None
-
-        target_label = f"workspace '{active_workspace.name}'"
-        media_lister, notes_lister, tag_lister = self._console_scope_picker_listers()
-
-        def _on_save(scope: Optional[RagScope]) -> None:
-            self.run_worker(
-                self._apply_console_workspace_scope_save(workspace_id, scope),
-                exclusive=True,
-                group="console-workspace-scope-save",
-            )
-
-        self.app.push_screen(
-            ConsoleScopePickerModal(
-                target_label,
-                None,
-                initial,
-                _on_save,
-                media_lister=media_lister,
-                notes_lister=notes_lister,
-                tag_lister=tag_lister,
-            )
-        )
-
-    async def _apply_console_workspace_scope_save(
-        self,
-        workspace_id: str,
-        scope: Optional[RagScope],
-    ) -> None:
-        """Persist (or clear) a workspace's RAG retrieval scope (task-13).
-
-        ``WorkspaceNotFound`` is caught deliberately -- the workspace may
-        have been archived/deleted (e.g. from Library > Workspaces) between
-        opening the picker and saving. Refreshes the ACTIVE Console
-        session's effective-scope display afterward: a workspace-scope
-        change can widen or narrow retrieval for every conversation linked
-        to it, but only the currently active session's row/chip are
-        mounted to refresh.
-
-        Args:
-            workspace_id: The workspace whose scope was just chosen.
-            scope: The new scope, or ``None`` to clear it.
-        """
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is None:
-            self.app_instance.notify(
-                "Couldn't save scope: workspace service is not ready.",
-                severity="error",
-            )
-            return
-        try:
-            await self._write_console_workspace_scope(
-                registry_service, workspace_id, scope
-            )
-        except WorkspaceNotFound:
-            logger.opt(exception=True).warning(
-                "Workspace scope save target missing: {}", workspace_id
-            )
-            self.app_instance.notify(
-                "Couldn't save scope: this workspace no longer exists.",
-                severity="error",
-            )
-            return
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Failed to write workspace scope for {}", workspace_id
-            )
-            self.app_instance.notify("Couldn't save workspace scope.", severity="error")
-            return
-        session = self._active_native_console_session()
-        if session is not None:
-            await self._refresh_console_effective_scope_and_sync(session)
 
     # Reactive property for sidebar state persistence
     sidebar_state = reactive({}, layout=False)
@@ -3225,15 +2851,96 @@ class ChatScreen(BaseAppScreen):
         self._agent_section_user_dismissed_while_busy = False
         self._console_rail_system_line_last: tuple[str, bool] | None = None
         self._console_rail_prune_dispatched = False
-        self._console_workspace_conversation_query = ""
-        self._console_workspace_conversation_search_timer: Any | None = None
-        self._console_workspace_conversation_search_token = 0
-        self._console_workspace_conversation_search_rows: tuple[
-            ConsoleWorkspaceConversationRow, ...
-        ] = ()
-        self._console_workspace_conversation_search_total: int | None = None
-        self._console_workspace_conversation_search_error = ""
-        self._console_workspace_conversation_workspace_id: str | None = None
+        #: Workspace policy context, lifecycle, and resume-flow state and
+        #: behaviour moved to `ConsoleWorkspaceController` (wave-2 console
+        #: decomposition, task 2) -- the largest cluster in wave 2.
+        #: `self._console_workspace_conversation_query`/`_search_timer`/
+        #: `_search_token`/`_search_rows`/`_search_total`/`_search_error`
+        #: stay readable/writable via the six proxy properties defined near
+        #: `_console_composer_or_none`, so `on_console_workspace_
+        #: conversation_search_changed` (a sibling conversation-browser
+        #: handler that only mirrors three of them) and `on_button_pressed`'s
+        #: workspace-search branches needed no change. See `workspace.py`'s
+        #: module docstring for the full map of what moved and why.
+        self._workspace = ConsoleWorkspaceController(
+            self,
+            app_instance=self.app_instance,
+            # Late-binding lambdas, not the bound methods directly -- same
+            # staleness reason as `ConsoleDictationController`'s own wiring
+            # just below: a bound method captured here would freeze the
+            # CURRENT screen method, invisible to a later
+            # `monkeypatch.setattr` on this screen instance.
+            chat_store_accessor=lambda: self._ensure_console_chat_store(),
+            current_chat_store_accessor=lambda: self._console_chat_store,
+            current_conversation_id_accessor=(
+                lambda: self._current_console_conversation_id()
+            ),
+            native_session_rows_accessor=(
+                lambda state: self._with_native_console_session_rows(state)
+            ),
+            conversation_browser_state_accessor=(
+                lambda state, current_conversation_id: (
+                    self._with_console_conversation_browser_state(
+                        state, current_conversation_id=current_conversation_id
+                    )
+                )
+            ),
+            capture_draft_switch_snapshot=(
+                lambda: self._capture_console_draft_switch_snapshot()
+            ),
+            sync_chat_core_state=lambda: self._sync_console_chat_core_state(),
+            sync_native_console_chat_ui=lambda: self._sync_native_console_chat_ui(),
+            sync_temporary_chip=lambda: self._sync_console_temporary_chip(),
+            default_session_settings_accessor=(
+                lambda: self._default_console_session_settings()
+            ),
+            scope_picker_listers_accessor=(
+                lambda: self._console_scope_picker_listers()
+            ),
+            active_native_session_accessor=(
+                lambda: self._active_native_console_session()
+            ),
+            refresh_effective_scope_and_sync=(
+                lambda session: self._refresh_console_effective_scope_and_sync(
+                    session
+                )
+            ),
+            messages_from_conversation_tree_accessor=(
+                lambda tree: self._console_messages_from_conversation_tree(tree)
+            ),
+            session_settings_for_resume_accessor=(
+                lambda conversation: self._console_session_settings_for_resume(
+                    conversation
+                )
+            ),
+            resolve_resumed_character_name=(
+                lambda character_id: self._resolve_resumed_character_name(
+                    character_id
+                )
+            ),
+            inject_resume_agent_markers_accessor=(
+                lambda messages, conversation_id: self._inject_resume_agent_markers(
+                    messages, conversation_id
+                )
+            ),
+            resolve_effective_scope_state=(
+                lambda session: self._resolve_console_effective_scope_state(session)
+            ),
+            sync_retrieval_scope_row=(
+                lambda: self._sync_console_retrieval_scope_row()
+            ),
+            note_follow_intent=lambda: self._note_console_follow_intent(),
+            focus_composer_if_needed=(
+                lambda **kwargs: self._focus_console_composer_if_needed(**kwargs)
+            ),
+            conversation_section_config_accessor=(
+                lambda: self._console_conversation_section_config()
+            ),
+            focus_conversation_search=(
+                lambda: self._focus_console_workspace_conversation_search()
+            ),
+            sync_workspace_context=lambda: self._sync_console_workspace_context(),
+        )
         self._console_conversation_browser_query = ""
         self._console_conversation_browser_search_timer: Any | None = None
         self._console_conversation_browser_search_token = 0
@@ -3846,7 +3553,7 @@ class ChatScreen(BaseAppScreen):
         store = self._ensure_console_chat_store()
         workspace_id = store.workspace_context.active_workspace_id
         session = store.ensure_session(
-            title=self._console_initial_session_title_for_workspace(workspace_id),
+            title=self._workspace._console_initial_session_title_for_workspace(workspace_id),
             workspace_id=workspace_id,
             settings=self._default_console_session_settings(),
         )
@@ -3927,7 +3634,7 @@ class ChatScreen(BaseAppScreen):
         store = self._ensure_console_chat_store()
         workspace_id = store.workspace_context.active_workspace_id
         session = store.ensure_session(
-            title=self._console_initial_session_title_for_workspace(workspace_id),
+            title=self._workspace._console_initial_session_title_for_workspace(workspace_id),
             workspace_id=workspace_id,
             settings=self._default_console_session_settings(),
         )
@@ -4090,7 +3797,7 @@ class ChatScreen(BaseAppScreen):
     ) -> ConsoleSettingsContextEstimate:
         """Return context usage for the active native Console settings snapshot."""
         settings = self._ensure_active_console_session_settings()
-        workspace_context = self._current_console_workspace_context()
+        workspace_context = self._workspace._current_console_workspace_context()
         staged_context_state = self._build_console_staged_context_state(
             self._pending_console_launch_context
         )
@@ -4647,121 +4354,6 @@ class ChatScreen(BaseAppScreen):
             group="console-sync",
         )
 
-    def _current_console_workspace_context(self) -> ConsoleWorkspaceContext:
-        """Return explicit workspace policy context for native Console sends."""
-        workspace_id = CONSOLE_GLOBAL_WORKSPACE_ID
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is not None:
-            try:
-                ensure_default_workspace = getattr(
-                    registry_service,
-                    "ensure_default_workspace",
-                    None,
-                )
-                active_workspace = (
-                    ensure_default_workspace()
-                    if callable(ensure_default_workspace)
-                    else registry_service.get_active_workspace()
-                )
-                candidate = getattr(active_workspace, "workspace_id", None)
-                if candidate:
-                    workspace_id = str(candidate)
-            except Exception:
-                logger.debug(
-                    "Console workspace registry was unavailable for send context"
-                )
-
-        staged_sources: list[ConsoleStagedSource] = []
-        pending_launch = self._pending_console_launch_context
-        if pending_launch is not None:
-            payload = pending_launch.payload
-            source_workspace = payload.get("workspace_id")
-            launch_workspace_id = (
-                str(source_workspace) if source_workspace else None
-            )
-            # RAG UX v2 PR-4: this builder is the SINGLE seam feeding both
-            # the Inspector rail badge ("{n} staged") and the settings
-            # context estimate, and it used to append exactly ONE staged
-            # source per launch -- so the status chip could read "Sources: 4
-            # staged" while its two siblings read 1. One row per bundle
-            # reference puts all three in the same vocabulary (and gives the
-            # workspace policy check a real per-source id to gate on).
-            bundle = evidence_bundle_from_launch(pending_launch)
-            references = bundle.references if bundle is not None else ()
-            for reference in references:
-                chunk_id = reference.metadata.get("chunk_id")
-                # Two chunks of one document share a source_id; qualify with
-                # the chunk id (exactly as the capture adapter does) so
-                # `allowed_sources`' source_id dedupe cannot collapse them.
-                staged_sources.append(
-                    ConsoleStagedSource(
-                        source_id=str(
-                            chunk_id
-                            if isinstance(chunk_id, str) and chunk_id
-                            else reference.source_id
-                        ),
-                        label=reference.title,
-                        source_type=reference.source_type,
-                        workspace_id=reference.workspace_id or launch_workspace_id,
-                    )
-                )
-            if not staged_sources:
-                # A launch with no (or an empty) evidence bundle is still one
-                # staged item -- the same fallback `console_staged_source_count`
-                # and the strip use, so all three surfaces still agree.
-                source_id = (
-                    payload.get("source_id")
-                    or payload.get("target_id")
-                    or payload.get("run_id")
-                    or pending_launch.title
-                )
-                staged_sources.append(
-                    ConsoleStagedSource(
-                        source_id=str(source_id),
-                        label=pending_launch.title,
-                        source_type=str(pending_launch.source),
-                        workspace_id=launch_workspace_id,
-                    )
-                )
-
-        return ConsoleWorkspaceContext(
-            active_workspace_id=workspace_id,
-            staged_sources=tuple(staged_sources),
-        )
-
-    def _active_console_workspace_id_for_conversation_search(self) -> str:
-        """Return the current active workspace id for Console conversation search."""
-        try:
-            workspace_id = str(
-                self._current_console_workspace_context().active_workspace_id or ""
-            ).strip()
-        except Exception:
-            logger.opt(exception=True).debug(
-                "Unable to read current workspace context for conversation search",
-            )
-            workspace_id = ""
-        if workspace_id:
-            return workspace_id
-        service = getattr(self.app_instance, "workspace_registry_service", None)
-        get_active_workspace = getattr(service, "get_active_workspace", None)
-        if callable(get_active_workspace):
-            try:
-                workspace = get_active_workspace()
-            except Exception:
-                logger.opt(exception=True).debug(
-                    "Unable to read active workspace for conversation search"
-                )
-                workspace = None
-            workspace_id = str(getattr(workspace, "workspace_id", "") or "").strip()
-            if workspace_id:
-                return workspace_id
-        store = self._console_chat_store
-        if store is not None and store.workspace_context.active_workspace_id:
-            return str(store.workspace_context.active_workspace_id)
-        return ""
-
     def _focus_console_workspace_conversation_search(self) -> None:
         """Restore focus to the conversation search input when it is mounted."""
         try:
@@ -4835,7 +4427,7 @@ class ChatScreen(BaseAppScreen):
             thinking_budget_tokens=selection_settings.thinking_budget_tokens,
             streaming=selection_settings.streaming,
             system_prompt=selection_settings.system_prompt,
-            workspace_context=self._current_console_workspace_context(),
+            workspace_context=self._workspace._current_console_workspace_context(),
         )
 
     def _active_console_provider_model_display(
@@ -4916,7 +4508,7 @@ class ChatScreen(BaseAppScreen):
                 )
             self._console_chat_store = ConsoleChatStore(
                 persistence=persistence,
-                workspace_context=self._current_console_workspace_context(),
+                workspace_context=self._workspace._current_console_workspace_context(),
                 on_scope_flushed=self._on_console_scope_flushed,
             )
         return self._console_chat_store
@@ -5545,111 +5137,6 @@ class ChatScreen(BaseAppScreen):
                 )
         return selection
 
-    def _activate_console_session_for_workspace(self, workspace_id: str) -> None:
-        """Activate or create the Console session for the selected workspace."""
-        target_workspace_id = str(workspace_id).strip()
-        if not target_workspace_id:
-            return
-        store = self._ensure_console_chat_store()
-        inherited_settings = None
-        if store.active_session_id is not None:
-            try:
-                inherited_settings = store.session_settings(store.active_session_id)
-            except KeyError:
-                inherited_settings = None
-        if store.active_session_id is not None:
-            for session in store.sessions():
-                if (
-                    session.id == store.active_session_id
-                    and session.workspace_id == target_workspace_id
-                ):
-                    return
-        for session in store.sessions():
-            if session.workspace_id == target_workspace_id:
-                self._capture_console_draft_switch_snapshot()
-                store.switch_session(session.id)
-                # task-7 review: this switches the active session with no
-                # other chip-refresh call anywhere in the caller chain (all
-                # three callers only run `_sync_native_console_chat_ui()`
-                # afterward, which never touches the temporary chip -- see
-                # `_sync_console_temporary_chip`). Without this the chip
-                # could keep reading "Temporary" on a workspace's saved
-                # session, or stay hidden on one that is actually temporary.
-                self._sync_console_temporary_chip()
-                return
-        self._capture_console_draft_switch_snapshot()
-        store.create_session(
-            title=self._console_workspace_session_title(target_workspace_id),
-            workspace_id=target_workspace_id,
-            settings=inherited_settings or self._default_console_session_settings(),
-        )
-        # task-7 review: `create_session` activates the new (never
-        # ephemeral -- no `ephemeral=` passed) session inline; same
-        # staleness risk as the switch branch above if the workspace's
-        # previous session was temporary.
-        self._sync_console_temporary_chip()
-
-    def _console_workspace_session_title(self, workspace_id: str) -> str:
-        """Return a readable title for an auto-created workspace Console tab."""
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        workspace_name = str(workspace_id).strip()
-        if registry_service is not None:
-            try:
-                workspace = registry_service.get_workspace(workspace_id)
-                if workspace is not None:
-                    workspace_name = workspace.name
-            except Exception:
-                logger.opt(exception=True).debug(
-                    "Unable to read Console workspace title"
-                )
-        if not workspace_name:
-            workspace_name = "Workspace"
-        return f"{workspace_name} Chat"
-
-    def _console_initial_session_title_for_workspace(
-        self, workspace_id: str | None
-    ) -> str:
-        """Return the first Console tab title for the active workspace."""
-        target_workspace_id = str(workspace_id or "").strip()
-        if not target_workspace_id or target_workspace_id in {
-            CONSOLE_GLOBAL_WORKSPACE_ID,
-            DEFAULT_WORKSPACE_ID,
-        }:
-            return DEFAULT_CONSOLE_SESSION_TITLE
-        return self._console_workspace_session_title(target_workspace_id)
-
-    def _set_active_workspace_for_console_session(self, session_id: str) -> None:
-        """Keep workspace context aligned when switching Console tabs."""
-        store = self._ensure_console_chat_store()
-        target_session = next(
-            (session for session in store.sessions() if session.id == session_id),
-            None,
-        )
-        if target_session is None:
-            return
-        workspace_id = str(target_session.workspace_id or "").strip()
-        if not workspace_id or workspace_id == CONSOLE_GLOBAL_WORKSPACE_ID:
-            return
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is None:
-            return
-        try:
-            active_workspace = registry_service.get_active_workspace()
-            if (
-                active_workspace is not None
-                and active_workspace.workspace_id == workspace_id
-            ):
-                return
-            registry_service.set_active_workspace(workspace_id)
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Unable to align Console workspace with selected tab",
-            )
-
     def _console_composer_or_none(self) -> ConsoleComposerBar | None:
         """Return the native Console composer when it is mounted."""
         composers = list(self.query("#console-native-composer"))
@@ -5741,6 +5228,71 @@ class ChatScreen(BaseAppScreen):
         availability` for the real implementation.
         """
         self._dictation._sync_console_dictation_availability()
+
+    # Workspace conversation-search state moved to `ConsoleWorkspaceController`
+    # (wave-2 console decomposition, task 2). These six properties keep
+    # `self._console_workspace_conversation_query`/`_search_timer`/
+    # `_search_token`/`_search_rows`/`_search_total`/`_search_error` readable
+    # AND writable exactly as before, for the two screen methods that are
+    # not part of the workspace cluster but still touch this state
+    # (`on_console_workspace_conversation_search_changed`, a sibling
+    # conversation-browser handler that mirrors three of them, and
+    # `on_button_pressed`'s workspace-search branches) and for tests that
+    # poke it directly -- each proxies straight through to `self._workspace`,
+    # so none of those call sites needed to change. `_console_workspace_
+    # conversation_workspace_id` needs no proxy: it is read and written only
+    # inside `_with_console_workspace_conversation_section`, itself moved.
+    @property
+    def _console_workspace_conversation_query(self) -> str:
+        return self._workspace._console_workspace_conversation_query
+
+    @_console_workspace_conversation_query.setter
+    def _console_workspace_conversation_query(self, value: str) -> None:
+        self._workspace._console_workspace_conversation_query = value
+
+    @property
+    def _console_workspace_conversation_search_timer(self) -> Any:
+        return self._workspace._console_workspace_conversation_search_timer
+
+    @_console_workspace_conversation_search_timer.setter
+    def _console_workspace_conversation_search_timer(self, value: Any) -> None:
+        self._workspace._console_workspace_conversation_search_timer = value
+
+    @property
+    def _console_workspace_conversation_search_token(self) -> int:
+        return self._workspace._console_workspace_conversation_search_token
+
+    @_console_workspace_conversation_search_token.setter
+    def _console_workspace_conversation_search_token(self, value: int) -> None:
+        self._workspace._console_workspace_conversation_search_token = value
+
+    @property
+    def _console_workspace_conversation_search_rows(
+        self,
+    ) -> tuple[ConsoleWorkspaceConversationRow, ...]:
+        return self._workspace._console_workspace_conversation_search_rows
+
+    @_console_workspace_conversation_search_rows.setter
+    def _console_workspace_conversation_search_rows(
+        self, value: tuple[ConsoleWorkspaceConversationRow, ...]
+    ) -> None:
+        self._workspace._console_workspace_conversation_search_rows = value
+
+    @property
+    def _console_workspace_conversation_search_total(self) -> int | None:
+        return self._workspace._console_workspace_conversation_search_total
+
+    @_console_workspace_conversation_search_total.setter
+    def _console_workspace_conversation_search_total(self, value: int | None) -> None:
+        self._workspace._console_workspace_conversation_search_total = value
+
+    @property
+    def _console_workspace_conversation_search_error(self) -> str:
+        return self._workspace._console_workspace_conversation_search_error
+
+    @_console_workspace_conversation_search_error.setter
+    def _console_workspace_conversation_search_error(self, value: str) -> None:
+        self._workspace._console_workspace_conversation_search_error = value
 
 
     def _speak_status(self, text: str) -> None:
@@ -9760,7 +9312,7 @@ class ChatScreen(BaseAppScreen):
         """
         store = self._ensure_console_chat_store()
         session = store.ensure_session(
-            title=self._console_initial_session_title_for_workspace(
+            title=self._workspace._console_initial_session_title_for_workspace(
                 store.workspace_context.active_workspace_id
             ),
             workspace_id=store.workspace_context.active_workspace_id,
@@ -10298,41 +9850,6 @@ class ChatScreen(BaseAppScreen):
                 write_conversation_scope, db, conversation_id, scope
             )
 
-    @staticmethod
-    async def _read_console_workspace_scope(
-        registry_service: Any, workspace_id: str
-    ) -> Optional[RagScope]:
-        """Read a workspace's stored scope, off-loop for a file-backed registry.
-
-        Mirrors ``_read_console_retrieval_scope``'s in-memory-DB guard
-        (task-13): ``LocalWorkspaceRegistryService.db`` is never actually
-        ``:memory:``-backed in production (``WorkspaceDB`` opens a brand-new
-        connection per call, incompatible with SQLite's ``:memory:``
-        semantics beyond a single call), but the guard is applied anyway
-        for the same defensive discipline the conversation-scope read uses.
-        """
-        db = getattr(registry_service, "db", None)
-        if getattr(db, "is_memory_db", False):
-            return registry_service.get_workspace_scope(workspace_id)
-        return await asyncio.to_thread(
-            registry_service.get_workspace_scope, workspace_id
-        )
-
-    @staticmethod
-    async def _write_console_workspace_scope(
-        registry_service: Any, workspace_id: str, scope: Optional[RagScope]
-    ) -> None:
-        """Write (or clear) a workspace's stored scope; see the read twin's
-        in-memory-registry-DB guard docstring for why this isn't
-        unconditionally ``asyncio.to_thread``."""
-        db = getattr(registry_service, "db", None)
-        if getattr(db, "is_memory_db", False):
-            registry_service.set_workspace_scope(workspace_id, scope)
-        else:
-            await asyncio.to_thread(
-                registry_service.set_workspace_scope, workspace_id, scope
-            )
-
     def _console_scope_picker_listers(
         self,
     ) -> tuple[Any, Any, Any]:
@@ -10383,7 +9900,7 @@ class ChatScreen(BaseAppScreen):
         )
         if workspace_id and registry_service is not None:
             try:
-                ws_scope = await self._read_console_workspace_scope(
+                ws_scope = await self._workspace._read_console_workspace_scope(
                     registry_service, workspace_id
                 )
             except Exception:
@@ -11266,27 +10783,6 @@ class ChatScreen(BaseAppScreen):
             logger.opt(exception=True).debug("avatar: pixels build failed")
             return Static("no avatar", id="console-character-avatar-empty")
 
-    def _console_session_id_for_workspace_conversation(
-        self,
-        conversation_id: str,
-    ) -> str | None:
-        """Return an open Console session id for a workspace conversation row."""
-        target = str(conversation_id or "").strip()
-        if not target:
-            return None
-        store = self._console_chat_store
-        if store is None:
-            return None
-        if target.startswith("native:"):
-            session_id = target.removeprefix("native:")
-            if any(session.id == session_id for session in store.sessions()):
-                return session_id
-            return None
-        for session in store.sessions():
-            if str(session.persisted_conversation_id or "") == target:
-                return session.id
-        return None
-
     @staticmethod
     def _console_message_role_from_persisted(
         message: dict[str, Any],
@@ -11588,212 +11084,6 @@ class ChatScreen(BaseAppScreen):
         broken.add(target)
         self._sync_console_workspace_context()
 
-    async def _resume_console_workspace_conversation(
-        self,
-        conversation_id: str,
-        *,
-        target_scope_type: str | None = None,
-        target_workspace_id: str | None = None,
-    ) -> bool | None:
-        """Load a persisted saved conversation into a native Console session.
-
-        Returns:
-            True on success; None on a transient failure this method already
-            notified about (service unavailable / load error); False when the
-            conversation record is missing - the caller owns that failure's
-            feedback (TASK-717).
-        """
-        target = str(conversation_id or "").strip()
-        if not target:
-            return None
-        # TASK-339: keystrokes typed while the conversation tree loads
-        # belong to the resumed session — snapshot the composer now.
-        self._capture_console_draft_switch_snapshot()
-        conversation_service = getattr(
-            self.app_instance,
-            "chat_conversation_scope_service",
-            None,
-        )
-        get_conversation_tree = getattr(
-            conversation_service, "get_conversation_tree", None
-        )
-        if not callable(get_conversation_tree):
-            self.app_instance.notify(
-                "Saved conversation resume is unavailable in this build.",
-                severity="warning",
-            )
-            return None
-
-        try:
-            # Task 8: raise the depth/root caps well past the service defaults
-            # (50) so a long or branchy conversation's full tree -- every
-            # branch, not just the latest -- is loaded intact, not truncated.
-            maybe_tree = get_conversation_tree(
-                target, mode="local", depth_cap=10_000, root_limit=10_000
-            )
-            tree = await maybe_tree if inspect.isawaitable(maybe_tree) else maybe_tree
-        except Exception:
-            logger.exception(
-                f"Unable to resume Console saved conversation: conversation_id={target}"
-            )
-            self.app_instance.notify(
-                "Unable to load this saved conversation.",
-                severity="error",
-            )
-            return None
-
-        if not isinstance(tree, dict) or not tree.get("conversation"):
-            # TASK-717: missing record - the caller owns this failure's UX
-            # (honest toast + marking the row visibly broken), so do not
-            # stack a second notification here.
-            return False
-
-        conversation = tree.get("conversation")
-        if not isinstance(conversation, dict):
-            conversation = {}
-        store = self._ensure_console_chat_store()
-        active_workspace_id = str(
-            store.workspace_context.active_workspace_id or ""
-        ).strip()
-        persisted_workspace_id = (
-            str(conversation.get("workspace_id")).strip()
-            if conversation.get("workspace_id") is not None
-            else ""
-        )
-        target_scope = str(target_scope_type or "").strip()
-        requested_workspace_id = (
-            str(target_workspace_id).strip() if target_workspace_id is not None else ""
-        )
-        if target_scope == "global":
-            workspace_id = CONSOLE_GLOBAL_WORKSPACE_ID
-        else:
-            workspace_id = (
-                persisted_workspace_id
-                or requested_workspace_id
-                or active_workspace_id
-                or None
-            )
-        title = str(conversation.get("title") or "Saved conversation").strip()
-        if not title:
-            title = "Saved conversation"
-        # Task 8: load the WHOLE persisted tree (every branch), then reconstruct
-        # the active branch from the stored active-leaf pointer. Loading all
-        # branches (not just the latest) is what makes off-path siblings
-        # navigable (swipe) right after resume.
-        all_nodes = self._console_messages_from_conversation_tree(tree)
-        db = getattr(self.app_instance, "chachanotes_db", None)
-        active_leaf_id = getattr(db, "get_conversation_active_leaf", lambda _c: None)(
-            target
-        )
-        raw_runtime_backend = conversation.get("runtime_backend")
-        if type(raw_runtime_backend) is str:
-            runtime_backend = raw_runtime_backend
-        else:
-            runtime_backend = ""
-        raw_assistant_kind = conversation.get("assistant_kind")
-        assistant_kind = raw_assistant_kind if type(raw_assistant_kind) is str else None
-        raw_assistant_id = conversation.get("assistant_id")
-        assistant_id = raw_assistant_id if type(raw_assistant_id) is str else None
-        raw_assistant_authority_id = conversation.get("assistant_authority_id")
-        assistant_authority_id = (
-            raw_assistant_authority_id
-            if type(raw_assistant_authority_id) is str
-            else None
-        )
-        raw_character_id = conversation.get("character_id")
-        character_id = (
-            raw_character_id
-            if (
-                runtime_backend == "local"
-                and assistant_kind == "character"
-                and type(raw_character_id) is int
-                and raw_character_id > 0
-                and assistant_id == str(raw_character_id)
-            )
-            else None
-        )
-        session = store.restore_persisted_session(
-            title=title,
-            workspace_id=workspace_id,
-            persisted_conversation_id=target,
-            all_nodes=all_nodes,
-            active_leaf_persisted_id=active_leaf_id,
-            settings=self._console_session_settings_for_resume(conversation),
-            runtime_backend=runtime_backend,
-            assistant_kind=assistant_kind,
-            assistant_id=assistant_id,
-            assistant_authority_id=assistant_authority_id,
-            character_id=character_id,
-        )
-        # Re-derive display-only agent TOOL markers from AgentRunsDB and overlay
-        # them onto the restored active-path VIEW (markers are never tree nodes;
-        # the next tree mutation's recompute rebuilds the view from live nodes
-        # and drops them, matching how live markers are ephemeral in Phase A).
-        store.apply_resume_marker_overlay(
-            session.id,
-            self._inject_resume_agent_markers(
-                store.messages_for_session(session.id), target
-            ),
-        )
-        # Local presentation remains keyed only by the numeric local
-        # projection. Opaque server identity never enters local card/avatar/
-        # dictionary lookup paths.
-        if (
-            session.runtime_backend == "local"
-            and session.assistant_kind == "character"
-            and session.character_id is not None
-        ):
-            character_name = await self._resolve_resumed_character_name(
-                session.character_id
-            )
-            if character_name:
-                session.character_name = character_name
-            # Always (re)set the label on a local character resume -- to the
-            # resolved name, or clear it when unresolved. ``settings`` are
-            # otherwise inherited from the currently active session, so
-            # leaving an inherited ``character_label`` in place would make
-            # a card-less resume show a *different* character's name.
-            if session.settings is not None:
-                session.settings = replace(
-                    session.settings, character_label=character_name
-                )
-        elif session.settings is not None:
-            session.settings = replace(session.settings, character_label="")
-        self._set_active_workspace_for_console_session(session.id)
-        # task-9/task-13: warm the EFFECTIVE (conversation ∩ workspace)
-        # scope cache for this session now (off-loop) so the Inspector row
-        # reflects reality immediately on resume, rather than defaulting to
-        # "everything" until the user opens Edit or saves a change (the
-        # picker's other two read triggers).
-        try:
-            await self._resolve_console_effective_scope_state(session)
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Failed to resolve retrieval scope for conversation {}", target
-            )
-        # task-10 review finding 2: warming the cache above is not enough
-        # by itself -- neither `_sync_native_console_chat_ui()` below nor
-        # its own `_sync_console_control_bar()` call ever touches the
-        # retrieval-scope row or `ConsoleStatusChips.sync_scope_chip`
-        # (`sync_scope_chip` is deliberately its own method, kept off the
-        # general control-bar sync tick -- see its docstring). Without this
-        # explicit call the MOUNTED row/chip stayed on whatever state they
-        # last rendered until the user opened Edit/Narrow or saved a
-        # change, even though the cache above already had the right
-        # answer. This is the same helper (and the same one-state,
-        # two-renderers push) the scope-picker save path already uses.
-        self._sync_console_retrieval_scope_row()
-        # Finding C: resuming a saved conversation switches the active
-        # conversation just as much as a tab switch does -- clear any
-        # sub-agent drill-in immediately rather than rely solely on the
-        # rail render path's defensive re-check on the next sync.
-        self._console_agent_drilldown_run_id = None
-        self._note_console_follow_intent()
-        self._sync_console_chat_core_state()
-        await self._sync_native_console_chat_ui()
-        self._focus_console_composer_if_needed(force=True)
-        return True
-
     def _on_console_scope_flushed(
         self, conversation_id: str, scope: Optional[RagScope]
     ) -> None:
@@ -11837,34 +11127,6 @@ class ChatScreen(BaseAppScreen):
                 exclusive=True,
                 group="console-effective-scope-refresh",
             )
-
-    def _build_console_workspace_context_state(self) -> ConsoleWorkspaceContextState:
-        current_conversation = self._current_console_conversation_id()
-        state = build_console_workspace_state(
-            registry_service=getattr(
-                self.app_instance, "workspace_registry_service", None
-            ),
-            current_conversation=current_conversation,
-            server_adapter_state=getattr(
-                self.app_instance,
-                "workspace_server_adapter_state",
-                None,
-            ),
-            acp_handoff_state=getattr(
-                self.app_instance,
-                "workspace_acp_handoff_state",
-                None,
-            ),
-        )
-        state = self._with_native_console_session_rows(state)
-        return self._with_console_conversation_browser_state(
-            state,
-            current_conversation_id=current_conversation,
-        )
-
-    @staticmethod
-    def _console_workspace_row_key(row: ConsoleWorkspaceConversationRow) -> str:
-        return str(row.conversation_id or "").strip()
 
     @staticmethod
     def _console_browser_row_key(row: ConsoleConversationBrowserInputRow) -> str:
@@ -11921,7 +11183,7 @@ class ChatScreen(BaseAppScreen):
         target_conversation_id = str(conversation_id or "").strip()
         if not target_row_key and not target_conversation_id:
             return None
-        state = self._build_console_workspace_context_state()
+        state = self._workspace._build_console_workspace_context_state()
         browser = state.conversation_browser
         if browser is None:
             return None
@@ -11951,47 +11213,6 @@ class ChatScreen(BaseAppScreen):
                         fallback = row
         return fallback
 
-    def _activate_console_workspace_for_browser_row(
-        self,
-        row: ConsoleConversationBrowserRow,
-    ) -> None:
-        """Align active workspace context before opening a browser row."""
-        scope_type = str(row.scope_type or "").strip()
-        if scope_type == "global":
-            return
-        workspace_id = str(row.workspace_id or "").strip()
-        if not workspace_id or workspace_id == CONSOLE_GLOBAL_WORKSPACE_ID:
-            return
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is None:
-            return
-        try:
-            active_workspace = registry_service.get_active_workspace()
-            if (
-                active_workspace is None
-                or active_workspace.workspace_id != workspace_id
-            ):
-                registry_service.set_active_workspace(workspace_id)
-                # TASK-713: opening a row from another workspace's group
-                # retargets the whole Console context; the Workspace status
-                # row is usually scrolled out of view at that moment, so the
-                # side effect needs an explicit announcement.
-                switched = registry_service.get_active_workspace()
-                switched_name = switched.name if switched is not None else workspace_id
-                self.app_instance.notify(
-                    f"Switched Console to {switched_name}.",
-                    severity="information",
-                )
-            self._ensure_console_chat_store().set_workspace_context(
-                self._current_console_workspace_context()
-            )
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Unable to activate Console workspace for browser row",
-            )
-
     def _console_session_id_for_browser_row(
         self,
         row: ConsoleConversationBrowserRow,
@@ -12007,7 +11228,7 @@ class ChatScreen(BaseAppScreen):
             return None
         row_key = str(row.row_key or "").strip()
         if row_key.startswith("native:"):
-            return self._console_session_id_for_workspace_conversation(row_key)
+            return self._workspace._console_session_id_for_workspace_conversation(row_key)
         conversation_id = str(row.conversation_id or "").strip()
         if not conversation_id:
             return None
@@ -12043,60 +11264,6 @@ class ChatScreen(BaseAppScreen):
             )
             return ("conversation", scope_type, workspace_id, conversation_id)
         return ("row", ChatScreen._console_browser_row_key(row))
-
-    def _console_browser_workspace_records(self) -> tuple[WorkspaceRecord, ...]:
-        """Return all local workspace records visible to the Console browser."""
-        service = getattr(self.app_instance, "workspace_registry_service", None)
-        if service is None:
-            return ()
-        ensure_default = getattr(service, "ensure_default_workspace", None)
-        if callable(ensure_default):
-            try:
-                ensure_default()
-            except Exception:
-                logger.opt(exception=True).debug(
-                    "Unable to ensure default workspace for Console browser"
-                )
-        list_workspaces = getattr(service, "list_workspaces", None)
-        if not callable(list_workspaces):
-            return ()
-        try:
-            return tuple(list_workspaces())
-        except Exception:
-            logger.opt(exception=True).debug(
-                "Unable to list Console browser workspaces"
-            )
-            return ()
-
-    def _console_browser_workspace_labels(self) -> dict[str, str]:
-        """Return workspace labels keyed by workspace id for browser rows."""
-        labels: dict[str, str] = {}
-        for record in self._console_browser_workspace_records():
-            workspace_id = str(record.workspace_id or "").strip()
-            if not workspace_id:
-                continue
-            labels[workspace_id] = (
-                "Chats"
-                if workspace_id == DEFAULT_WORKSPACE_ID
-                else str(record.name or workspace_id)
-            )
-        labels.setdefault(DEFAULT_WORKSPACE_ID, "Chats")
-        return labels
-
-    def _console_browser_workspace_label(
-        self,
-        workspace_id: str | None,
-        labels: dict[str, str] | None = None,
-    ) -> str:
-        """Return display label for a workspace/global browser row."""
-        if not workspace_id or workspace_id == CONSOLE_GLOBAL_WORKSPACE_ID:
-            return "Chats"
-        if workspace_id == DEFAULT_WORKSPACE_ID:
-            return "Chats"
-        workspace_labels = (
-            labels if labels is not None else self._console_browser_workspace_labels()
-        )
-        return workspace_labels.get(workspace_id, workspace_id)
 
     def _starred_console_conversation_ids(self) -> set[str]:
         """Return locally starred durable conversation ids."""
@@ -12140,7 +11307,7 @@ class ChatScreen(BaseAppScreen):
         store = self._console_chat_store
         if store is None:
             return []
-        labels = self._console_browser_workspace_labels()
+        labels = self._workspace._console_browser_workspace_labels()
         starred_ids = self._starred_console_conversation_ids()
         active_session_id = store.active_session_id
         controller = getattr(self, "_console_chat_controller", None)
@@ -12177,7 +11344,7 @@ class ChatScreen(BaseAppScreen):
                 title=str(session.title or "Untitled conversation"),
                 scope_type=scope_type,
                 workspace_id=workspace_id,
-                workspace_label=self._console_browser_workspace_label(
+                workspace_label=self._workspace._console_browser_workspace_label(
                     workspace_id, labels
                 ),
                 status="active session" if selected else "open session",
@@ -12198,7 +11365,7 @@ class ChatScreen(BaseAppScreen):
         list_conversations = getattr(service, "list_workspace_conversations", None)
         if not callable(list_conversations):
             return []
-        labels = self._console_browser_workspace_labels()
+        labels = self._workspace._console_browser_workspace_labels()
         starred_ids = self._starred_console_conversation_ids()
         current_conversation = (
             current_conversation_id or self._current_console_conversation_id()
@@ -12208,11 +11375,11 @@ class ChatScreen(BaseAppScreen):
             str(active_session.workspace_id or "").strip()
             if active_session is not None
             else str(
-                self._current_console_workspace_context().active_workspace_id or ""
+                self._workspace._current_console_workspace_context().active_workspace_id or ""
             ).strip()
         )
         rows: list[ConsoleConversationBrowserInputRow] = []
-        for record in self._console_browser_workspace_records():
+        for record in self._workspace._console_browser_workspace_records():
             workspace_id = str(record.workspace_id or "").strip()
             if not workspace_id:
                 continue
@@ -12237,7 +11404,7 @@ class ChatScreen(BaseAppScreen):
                     title=title,
                     scope_type="workspace",
                     workspace_id=workspace_id,
-                    workspace_label=self._console_browser_workspace_label(
+                    workspace_label=self._workspace._console_browser_workspace_label(
                         workspace_id, labels
                     ),
                     status=str(getattr(membership, "role", "") or "workspace-thread"),
@@ -12282,11 +11449,11 @@ class ChatScreen(BaseAppScreen):
         if not services:
             return [], None, ""
 
-        labels = self._console_browser_workspace_labels()
+        labels = self._workspace._console_browser_workspace_labels()
         scopes: list[tuple[str, str | None]] = [("global", None)]
         scopes.extend(
             ("workspace", str(record.workspace_id))
-            for record in self._console_browser_workspace_records()
+            for record in self._workspace._console_browser_workspace_records()
             if str(record.workspace_id or "").strip()
         )
         last_error = ""
@@ -12396,7 +11563,7 @@ class ChatScreen(BaseAppScreen):
                         title=str(item.get("title") or "Untitled conversation"),
                         scope_type=item_scope_type,
                         workspace_id=normalized_workspace_id,
-                        workspace_label=self._console_browser_workspace_label(
+                        workspace_label=self._workspace._console_browser_workspace_label(
                             normalized_workspace_id,
                             labels,
                         ),
@@ -12493,11 +11660,11 @@ class ChatScreen(BaseAppScreen):
         if not services:
             return [], None, ""
 
-        labels = self._console_browser_workspace_labels()
+        labels = self._workspace._console_browser_workspace_labels()
         scopes: list[tuple[str, str | None]] = [("global", None)]
         scopes.extend(
             ("workspace", str(record.workspace_id))
-            for record in self._console_browser_workspace_records()
+            for record in self._workspace._console_browser_workspace_records()
             if str(record.workspace_id or "").strip()
         )
         last_error = ""
@@ -12593,7 +11760,7 @@ class ChatScreen(BaseAppScreen):
                         title=str(item.get("title") or "Untitled conversation"),
                         scope_type=item_scope_type,
                         workspace_id=normalized_workspace_id,
-                        workspace_label=self._console_browser_workspace_label(
+                        workspace_label=self._workspace._console_browser_workspace_label(
                             normalized_workspace_id,
                             labels,
                         ),
@@ -12660,286 +11827,6 @@ class ChatScreen(BaseAppScreen):
         else:
             total = None
         return rows, total, self._console_conversation_browser_error or sync_error
-
-    def _selected_console_workspace_conversation_summary(
-        self,
-        rows: list[ConsoleWorkspaceConversationRow],
-    ) -> str:
-        selected = next((row for row in rows if row.selected), None)
-        if selected is None:
-            return "No active conversation."
-        title = ConsoleWorkspaceContextTray._conversation_title(selected.title)
-        detail = ConsoleWorkspaceContextTray._conversation_detail_status(
-            selected.status
-        )
-        return f"{title} - {detail or 'conversation'}"
-
-    def _merge_console_workspace_rows(
-        self,
-        primary: list[ConsoleWorkspaceConversationRow],
-        secondary: list[ConsoleWorkspaceConversationRow],
-    ) -> list[ConsoleWorkspaceConversationRow]:
-        merged: list[ConsoleWorkspaceConversationRow] = []
-        seen: set[str] = set()
-        for row in primary + secondary:
-            key = self._console_workspace_row_key(row)
-            if not key or key in seen:
-                continue
-            seen.add(key)
-            merged.append(row)
-        return merged
-
-    def _native_console_rows_for_workspace_search(
-        self,
-        workspace_id: str,
-        query: str,
-    ) -> list[ConsoleWorkspaceConversationRow]:
-        """Return matching open native sessions for the active workspace search."""
-        store = self._console_chat_store
-        if store is None:
-            return []
-        needle = str(query or "").strip().lower()
-        rows: list[ConsoleWorkspaceConversationRow] = []
-        active_session_id = store.active_session_id
-        for session in store.sessions():
-            selected = session.id == active_session_id
-            session_workspace_id = str(session.workspace_id or "").strip()
-            if (
-                workspace_id
-                and workspace_id != CONSOLE_GLOBAL_WORKSPACE_ID
-                and session_workspace_id != workspace_id
-                and not selected
-            ):
-                continue
-            title = str(session.title or "Untitled conversation")
-            if needle and needle not in title.lower():
-                continue
-            conversation_id = (
-                str(session.persisted_conversation_id)
-                if session.persisted_conversation_id
-                else f"native:{session.id}"
-            )
-            rows.append(
-                ConsoleWorkspaceConversationRow(
-                    conversation_id=conversation_id,
-                    title=title,
-                    status="active" if selected else "open",
-                    selected=selected,
-                )
-            )
-        return rows
-
-    def _membership_console_rows_for_workspace_search(
-        self,
-        workspace_id: str,
-        query: str,
-    ) -> list[ConsoleWorkspaceConversationRow]:
-        """Return matching workspace conversation membership rows."""
-        service = getattr(self.app_instance, "workspace_registry_service", None)
-        list_conversations = getattr(service, "list_workspace_conversations", None)
-        if not callable(list_conversations) or not workspace_id:
-            return []
-        needle = str(query or "").strip().lower()
-        try:
-            memberships = list_conversations(workspace_id)
-        except Exception:
-            logger.opt(exception=True).debug(
-                "Unable to search workspace conversation memberships"
-            )
-            return []
-        rows: list[ConsoleWorkspaceConversationRow] = []
-        current_conversation = self._current_console_conversation_id()
-        for membership in memberships:
-            title = str(
-                getattr(membership, "title", "") or getattr(membership, "item_id", "")
-            )
-            if needle and needle not in title.lower():
-                continue
-            conversation_id = str(getattr(membership, "item_id", "") or "")
-            rows.append(
-                ConsoleWorkspaceConversationRow(
-                    conversation_id=conversation_id,
-                    title=title,
-                    status=str(getattr(membership, "role", "") or "workspace-thread"),
-                    selected=bool(
-                        current_conversation and conversation_id == current_conversation
-                    ),
-                )
-            )
-        return rows
-
-    async def _persisted_console_rows_for_workspace_search(
-        self,
-        workspace_id: str,
-        query: str,
-    ) -> tuple[list[ConsoleWorkspaceConversationRow], int | None, str]:
-        """Return persisted workspace conversation search rows, total, and error copy."""
-        scope_service = getattr(
-            self.app_instance,
-            "chat_conversation_scope_service",
-            None,
-        )
-        list_conversations = getattr(scope_service, "list_conversations", None)
-        if not callable(list_conversations) or not workspace_id:
-            return [], None, ""
-        if (
-            hasattr(scope_service, "local_service")
-            and getattr(scope_service, "local_service", None) is None
-        ):
-            return [], None, ""
-        try:
-            result = list_conversations(
-                mode="local",
-                query=query,
-                scope_type="workspace",
-                workspace_id=workspace_id,
-                limit=CONSOLE_WORKSPACE_CONVERSATION_RESULT_LIMIT,
-                offset=0,
-            )
-            result = await result if inspect.isawaitable(result) else result
-        except Exception as exc:
-            if (
-                isinstance(exc, ValueError)
-                and "service is unavailable" in str(exc).lower()
-            ):
-                logger.debug(
-                    "Local persisted conversation search service is unavailable"
-                )
-                return [], None, ""
-            logger.exception("Unable to search Console workspace conversations")
-            return [], None, "Workspace conversation search is unavailable."
-        if not isinstance(result, dict):
-            return [], 0, ""
-        items = result.get("items")
-        if not isinstance(items, list):
-            items = []
-        total = result.get("total")
-        if total is None:
-            pagination = result.get("pagination")
-            if isinstance(pagination, dict):
-                total = pagination.get("total")
-        try:
-            total_count = int(total)
-        except (TypeError, ValueError):
-            total_count = len(items)
-        current_conversation = self._current_console_conversation_id()
-        rows: list[ConsoleWorkspaceConversationRow] = []
-        for item in items:
-            if not isinstance(item, dict):
-                continue
-            conversation_id = str(item.get("id") or "").strip()
-            if not conversation_id:
-                continue
-            rows.append(
-                ConsoleWorkspaceConversationRow(
-                    conversation_id=conversation_id,
-                    title=str(item.get("title") or "Untitled conversation"),
-                    status=str(item.get("state") or "workspace-thread"),
-                    selected=bool(
-                        current_conversation and current_conversation == conversation_id
-                    ),
-                )
-            )
-        return rows, total_count, ""
-
-    def _refresh_console_workspace_conversation_search_if_current(
-        self,
-        workspace_id: str,
-        query: str,
-        token: int,
-        *,
-        restore_focus: bool = False,
-    ) -> bool:
-        """Refresh search results when workspace, query, and token still match."""
-        if token != self._console_workspace_conversation_search_token:
-            return False
-        if workspace_id != self._active_console_workspace_id_for_conversation_search():
-            return False
-        if query != self._console_workspace_conversation_query:
-            return False
-        self._sync_console_workspace_context()
-        if restore_focus:
-            self.call_after_refresh(self._focus_console_workspace_conversation_search)
-        return True
-
-    async def _refresh_console_workspace_conversation_search(
-        self,
-        workspace_id: str,
-        query: str,
-        token: int,
-    ) -> None:
-        """Refresh search results only if workspace and query are still current."""
-        if token != self._console_workspace_conversation_search_token:
-            return
-        if workspace_id != self._active_console_workspace_id_for_conversation_search():
-            return
-        if query != self._console_workspace_conversation_query:
-            return
-        if not str(query or "").strip():
-            self._console_workspace_conversation_search_rows = ()
-            self._console_workspace_conversation_search_total = None
-            self._console_workspace_conversation_search_error = ""
-            self._sync_console_workspace_context()
-            self.call_after_refresh(self._focus_console_workspace_conversation_search)
-            return
-        native_rows = self._native_console_rows_for_workspace_search(
-            workspace_id,
-            query,
-        )
-        membership_rows = self._membership_console_rows_for_workspace_search(
-            workspace_id,
-            query,
-        )
-        local_rows = self._merge_console_workspace_rows(native_rows, membership_rows)
-        self._console_workspace_conversation_search_rows = tuple(local_rows)
-        self._console_workspace_conversation_search_total = len(local_rows)
-        self._console_workspace_conversation_search_error = ""
-        self._sync_console_workspace_context()
-        self.call_after_refresh(self._focus_console_workspace_conversation_search)
-        (
-            persisted_rows,
-            persisted_total,
-            error_copy,
-        ) = await self._persisted_console_rows_for_workspace_search(
-            workspace_id,
-            query,
-        )
-        if token != self._console_workspace_conversation_search_token:
-            return
-        if workspace_id != self._active_console_workspace_id_for_conversation_search():
-            return
-        if query != self._console_workspace_conversation_query:
-            return
-        merged = self._merge_console_workspace_rows(
-            local_rows,
-            persisted_rows,
-        )
-        result_total = persisted_total
-        if result_total is None or result_total < len(merged):
-            result_total = len(merged)
-        self._console_workspace_conversation_search_rows = tuple(merged)
-        self._console_workspace_conversation_search_total = result_total
-        self._console_workspace_conversation_search_error = error_copy
-        self._sync_console_workspace_context()
-        self.call_after_refresh(self._focus_console_workspace_conversation_search)
-
-    async def _refresh_console_workspace_conversation_search_after_selection(
-        self,
-    ) -> None:
-        """Refresh active search rows after a conversation row changes selection."""
-        query = self._console_workspace_conversation_query
-        if not query.strip():
-            return
-        if self._console_workspace_conversation_search_timer is not None:
-            self._console_workspace_conversation_search_timer.stop()
-            self._console_workspace_conversation_search_timer = None
-        self._console_workspace_conversation_search_token += 1
-        token = self._console_workspace_conversation_search_token
-        await self._refresh_console_workspace_conversation_search(
-            self._active_console_workspace_id_for_conversation_search(),
-            query,
-            token,
-        )
 
     async def _refresh_console_conversation_browser_search(
         self,
@@ -13067,75 +11954,9 @@ class ChatScreen(BaseAppScreen):
         return replace(
             state,
             conversation_rows=tuple(
-                self._merge_console_workspace_rows(native_rows, rows)
+                self._workspace._merge_console_workspace_rows(native_rows, rows)
             ),
         )
-
-    def _with_console_workspace_conversation_section(
-        self,
-        state: ConsoleWorkspaceContextState,
-    ) -> ConsoleWorkspaceContextState:
-        """Attach renderable Conversations subsection state to workspace context."""
-        workspace_id = ""
-        try:
-            workspace_id = str(
-                self._current_console_workspace_context().active_workspace_id or ""
-            ).strip()
-        except Exception:
-            workspace_id = ""
-        store = self._console_chat_store
-        if not workspace_id:
-            if store is not None and store.workspace_context.active_workspace_id:
-                workspace_id = str(store.workspace_context.active_workspace_id)
-            elif state.workspace_label.startswith("Workspace: "):
-                workspace_id = state.workspace_label.removeprefix("Workspace: ").strip()
-
-        if self._console_workspace_conversation_workspace_id != workspace_id:
-            self._console_workspace_conversation_query = ""
-            self._console_workspace_conversation_search_token += 1
-            self._console_workspace_conversation_search_rows = ()
-            self._console_workspace_conversation_search_total = None
-            self._console_workspace_conversation_search_error = ""
-            self._console_workspace_conversation_workspace_id = workspace_id
-
-        rows = list(state.conversation_rows)
-        if self._console_workspace_conversation_query.strip():
-            rows = list(self._console_workspace_conversation_search_rows)
-        selected_summary = self._selected_console_workspace_conversation_summary(rows)
-        query = self._console_workspace_conversation_query
-        result_total = (
-            self._console_workspace_conversation_search_total if query.strip() else None
-        )
-        if (
-            query.strip()
-            and result_total is None
-            and not self._console_workspace_conversation_search_error
-        ):
-            result_total = len(rows)
-        status_copy = console_workspace_conversation_result_copy(
-            query=query,
-            result_total_count=result_total,
-            result_limit=CONSOLE_WORKSPACE_CONVERSATION_RESULT_LIMIT,
-        )
-        section = ConsoleWorkspaceConversationSectionState(
-            workspace_id=workspace_id,
-            collapsed=self._console_workspace_conversations_collapsed(workspace_id),
-            query=query,
-            selected_summary=selected_summary,
-            rows=tuple(rows),
-            workspace_total_count=len(rows),
-            result_total_count=result_total,
-            status_copy=status_copy,
-            empty_copy=(
-                "No matches in this workspace."
-                if query.strip()
-                else state.conversation_empty_copy
-            ),
-            search_enabled=True,
-            new_conversation_enabled=state.new_conversation_enabled,
-            error_copy=self._console_workspace_conversation_search_error,
-        )
-        return replace(state, conversation_section=section)
 
     def _console_subagent_counts_refresh_needed(self, row_ids: frozenset) -> bool:
         """Decide whether the sub-agent badge-count cache needs a DB round trip.
@@ -13230,7 +12051,7 @@ class ChatScreen(BaseAppScreen):
         subagent_counts = self._console_subagent_counts_for_rows(bridge, rows)
         browser = build_console_conversation_browser_state(
             rows=rows,
-            active_workspace_id=self._current_console_workspace_context().active_workspace_id,
+            active_workspace_id=self._workspace._current_console_workspace_context().active_workspace_id,
             group_collapse_preferences=(
                 self._console_conversation_browser_collapse_preferences()
             ),
@@ -13241,7 +12062,7 @@ class ChatScreen(BaseAppScreen):
             result_limit=CONSOLE_CONVERSATION_BROWSER_RESULT_LIMIT,
             subagent_counts=subagent_counts,
         )
-        legacy_state = self._with_console_workspace_conversation_section(state)
+        legacy_state = self._workspace._with_console_workspace_conversation_section(state)
         return replace(
             state,
             conversation_browser=browser,
@@ -13315,34 +12136,6 @@ class ChatScreen(BaseAppScreen):
             collapsed_groups = {}
             browser_config["collapsed_groups"] = collapsed_groups
         collapsed_groups[normalized_group_id] = bool(collapsed)
-
-    def _console_workspace_conversations_collapsed(
-        self,
-        workspace_id: str | None,
-    ) -> bool:
-        """Return stored collapse preference for one workspace."""
-        key = str(workspace_id or "global").strip() or "global"
-        app_config = getattr(self.app_instance, "app_config", None)
-        if not isinstance(app_config, dict):
-            return False
-        console_config = app_config.get("console")
-        if not isinstance(console_config, dict):
-            return False
-        section_config = console_config.get("conversation_section")
-        if not isinstance(section_config, dict):
-            return False
-        value = section_config.get(key)
-        return bool(value.get("collapsed")) if isinstance(value, dict) else False
-
-    def _set_console_workspace_conversations_collapsed(
-        self,
-        workspace_id: str | None,
-        collapsed: bool,
-    ) -> None:
-        """Store collapse preference for one workspace in memory."""
-        key = str(workspace_id or "global").strip() or "global"
-        section_config = self._console_conversation_section_config()
-        section_config[key] = {"collapsed": bool(collapsed)}
 
     def _console_rail_state_config(self) -> dict[str, Any]:
         """Return mutable Console rail-state config, initializing it if needed."""
@@ -13785,7 +12578,7 @@ class ChatScreen(BaseAppScreen):
         workspace_context_state: ConsoleWorkspaceContextState,
     ) -> ConsoleRailState:
         """Build the effective Console rail state for the current composition."""
-        workspace_context = self._current_console_workspace_context()
+        workspace_context = self._workspace._current_console_workspace_context()
         active_session_id = (
             self._console_chat_store.active_session_id
             if self._console_chat_store is not None
@@ -13970,7 +12763,7 @@ class ChatScreen(BaseAppScreen):
         pending_launch = self._pending_console_launch_context
         staged_context_state = self._build_console_staged_context_state(pending_launch)
         inspector_state = self._build_console_inspector_state(pending_launch)
-        workspace_context_state = self._build_console_workspace_context_state()
+        workspace_context_state = self._workspace._build_console_workspace_context_state()
         rail_state = self._build_console_rail_state(
             staged_context_state=staged_context_state,
             inspector_state=inspector_state,
@@ -14047,7 +12840,7 @@ class ChatScreen(BaseAppScreen):
         notify_on_failure: bool = True,
     ) -> ConsoleRailState:
         """Persist requested Console rail preference changes and return new state."""
-        workspace_context = self._current_console_workspace_context()
+        workspace_context = self._workspace._current_console_workspace_context()
         preference_key = build_console_rail_preference_key(
             workspace_id=workspace_context.active_workspace_id,
             conversation_id=self._current_console_rail_conversation_id(),
@@ -14130,7 +12923,7 @@ class ChatScreen(BaseAppScreen):
                 "#console-workspace-context",
                 ConsoleWorkspaceContextTray,
             )
-            state = self._build_console_workspace_context_state()
+            state = self._workspace._build_console_workspace_context_state()
             # TASK-251: read the widget's OWN current state before it's
             # overwritten -- this is intentionally not a screen-level cache
             # (which would go stale across a full-screen recompose); the
@@ -14213,7 +13006,7 @@ class ChatScreen(BaseAppScreen):
         except (NoMatches, QueryError):
             return
 
-        state = self._build_console_workspace_context_state()
+        state = self._workspace._build_console_workspace_context_state()
 
         if not self.query("#console-new-workspace-conversation"):
             new_button = Button(
@@ -15022,7 +13815,7 @@ class ChatScreen(BaseAppScreen):
                 ConsoleDisplayRow("Conversation source", "none"),
             )
 
-        workspace_state = self._build_console_workspace_context_state()
+        workspace_state = self._workspace._build_console_workspace_context_state()
         workspace_label = str(workspace_state.workspace_label or "").strip()
         if workspace_label.startswith("Workspace: "):
             workspace_label = workspace_label.removeprefix("Workspace: ").strip()
@@ -16166,7 +14959,7 @@ class ChatScreen(BaseAppScreen):
         control_state = self._build_console_control_state(pending_launch)
         staged_context_state = self._build_console_staged_context_state(pending_launch)
         inspector_state = self._build_console_inspector_state(pending_launch)
-        workspace_context_state = self._build_console_workspace_context_state()
+        workspace_context_state = self._workspace._build_console_workspace_context_state()
         # task-10: built once, shared verbatim by the header's "Scope" chip
         # and the Inspector's retrieval-scope row below -- one zero-DB
         # state, two renderers.
@@ -17778,7 +16571,7 @@ class ChatScreen(BaseAppScreen):
                 store.set_session_draft(session.id, suggested_prompt)
             else:
                 session = store.ensure_session(
-                    title=self._console_initial_session_title_for_workspace(
+                    title=self._workspace._console_initial_session_title_for_workspace(
                         store.workspace_context.active_workspace_id
                     ),
                     workspace_id=store.workspace_context.active_workspace_id,
@@ -18439,7 +17232,7 @@ class ChatScreen(BaseAppScreen):
             return
         store = self._ensure_console_chat_store()
         store.ensure_session(
-            title=self._console_initial_session_title_for_workspace(
+            title=self._workspace._console_initial_session_title_for_workspace(
                 store.workspace_context.active_workspace_id
             ),
             workspace_id=store.workspace_context.active_workspace_id,
@@ -18509,7 +17302,7 @@ class ChatScreen(BaseAppScreen):
                 pass
         else:
             session = store.ensure_session(
-                title=self._console_initial_session_title_for_workspace(
+                title=self._workspace._console_initial_session_title_for_workspace(
                     store.workspace_context.active_workspace_id
                 ),
                 workspace_id=store.workspace_context.active_workspace_id,
@@ -23417,7 +22210,7 @@ class ChatScreen(BaseAppScreen):
                 and last_round_ids <= self._console_toasted_park_round_ids
             ):
                 return
-        session_title, workspace_name = self._console_session_title_and_workspace_name(
+        session_title, workspace_name = self._workspace._console_session_title_and_workspace_name(
             controller, session_id
         )
         self.app_instance.notify(
@@ -23468,46 +22261,6 @@ class ChatScreen(BaseAppScreen):
                 ids.add(f"script:{request_id}")
         return frozenset(ids)
 
-    def _console_session_title_and_workspace_name(
-        self, controller: ConsoleChatController, session_id: str
-    ) -> tuple[str, str]:
-        """Return ``(session_title, workspace_name)`` for a fleet toast.
-
-        Fix wave (rider 6, final review): shared by ``_park_console_
-        approval`` and ``_notify_console_run_outcome``, which previously
-        duplicated this exact lookup byte-for-byte. Falls back to the raw
-        ``session_id``/workspace id when the session has already closed or
-        the workspace can't be resolved -- a toast about a session that
-        vanished microseconds ago must still say SOMETHING coherent rather
-        than raise.
-        """
-        session_title = session_id
-        workspace_id = CONSOLE_GLOBAL_WORKSPACE_ID
-        for session in controller.store.sessions():
-            if session.id == session_id:
-                session_title = session.title
-                workspace_id = session.workspace_id
-                break
-        return session_title, self._console_workspace_display_name(workspace_id)
-
-    def _console_workspace_display_name(self, workspace_id: str) -> str:
-        """Return ``workspace_id``'s display name via the registry, falling
-        back to the raw id when the service is unavailable or the
-        workspace can't be resolved (PA-T9 toast copy)."""
-        registry_service = getattr(
-            self.app_instance, "workspace_registry_service", None
-        )
-        if registry_service is not None:
-            try:
-                workspace = registry_service.get_workspace(workspace_id)
-                if workspace is not None and workspace.name:
-                    return workspace.name
-            except Exception:
-                logger.opt(exception=True).debug(
-                    "Unable to resolve Console workspace name for approval toast"
-                )
-        return str(workspace_id)
-
     def _notify_console_run_outcome(
         self, session_id: str, status: ConsoleRunStatus
     ) -> None:
@@ -23537,7 +22290,7 @@ class ChatScreen(BaseAppScreen):
         controller = self._console_chat_controller
         if controller is None:
             return
-        session_title, workspace_name = self._console_session_title_and_workspace_name(
+        session_title, workspace_name = self._workspace._console_session_title_and_workspace_name(
             controller, session_id
         )
         verb = "finished" if status is ConsoleRunStatus.COMPLETED else "failed"
@@ -23710,7 +22463,7 @@ class ChatScreen(BaseAppScreen):
         ):
             event.stop()
             group_id = str(getattr(event.button, "group_id", "") or "").strip()
-            state = self._build_console_workspace_context_state()
+            state = self._workspace._build_console_workspace_context_state()
             section_id = group_id.removeprefix("section:")
             section = None
             browser = state.conversation_browser
@@ -23732,7 +22485,7 @@ class ChatScreen(BaseAppScreen):
         ):
             event.stop()
             group_id = str(getattr(event.button, "group_id", "") or "").strip()
-            state = self._build_console_workspace_context_state()
+            state = self._workspace._build_console_workspace_context_state()
             group = None
             browser = state.conversation_browser
             if browser is not None:
@@ -23835,7 +22588,7 @@ class ChatScreen(BaseAppScreen):
                 if section is None:
                     return
                 collapsed = not bool(section.collapsed)
-                self._set_console_workspace_conversations_collapsed(
+                self._workspace._set_console_workspace_conversations_collapsed(
                     section.workspace_id,
                     collapsed,
                 )
@@ -23887,12 +22640,12 @@ class ChatScreen(BaseAppScreen):
                 conversation_id=conversation_id,
             )
             if browser_row is not None:
-                self._activate_console_workspace_for_browser_row(browser_row)
+                self._workspace._activate_console_workspace_for_browser_row(browser_row)
                 row_conversation_id = str(browser_row.conversation_id or "").strip()
                 session_id = self._console_session_id_for_browser_row(browser_row)
             else:
                 row_conversation_id = conversation_id
-                session_id = self._console_session_id_for_workspace_conversation(
+                session_id = self._workspace._console_session_id_for_workspace_conversation(
                     conversation_id
                 )
             if session_id is None:
@@ -23909,7 +22662,7 @@ class ChatScreen(BaseAppScreen):
                 # flag; the finally covers the not-resumable/error return).
                 self._set_console_conversation_row_loading(row_conversation_id, True)
                 try:
-                    resumed = await self._resume_console_workspace_conversation(
+                    resumed = await self._workspace._resume_console_workspace_conversation(
                         row_conversation_id,
                         target_scope_type=(
                             browser_row.scope_type if browser_row is not None else None
@@ -23943,7 +22696,7 @@ class ChatScreen(BaseAppScreen):
             controller = self._ensure_console_chat_controller()
             if controller.store.active_session_id != session_id:
                 if browser_row is None:
-                    self._set_active_workspace_for_console_session(session_id)
+                    self._workspace._set_active_workspace_for_console_session(session_id)
                 controller.switch_session(session_id)
                 await self._sync_native_console_chat_ui()
                 # task-7 review: an already-open native tab for this
